148 Since Plan
Question: 354
Title: Russian Doll Envelopes
Difficulty: Hard
Beat: 72%
Date: 2019-05-30
Comments: Very hard problem. First, I need to generate a sorted
subquence so that env[0] is increasing and env[1] is descreasing. Then I
need to find out the longest subsequence if [env[1] for env in envs].

For finding the longest increasing subquence, I could use a DP solution
which is the same as in wikipedia.